why use an event log when you can manually toggle bool variables
