+++ index.js
@@ -28,6 +28,16 @@ process.on("warning", (error) => {
 
 client.login(process.env.DISCORD_TOKEN);
 
+var isFirstStartEvent = true;
+
+function toggleFirstStartTrue() {
+  isFirstStartEvent = true;
+}
+
+function toggleFirstStartFalse() {
+  isFirstStartEvent = false;
+}
+
 const app = express();
 const port = process.env.API_PORT;
 const frontendUrl = process.env.FRONTEND_URL;
@@ -43,3 +53,5 @@ app.use(errorHandler);
 app.listen(port, () => {
   console.log(`API is accessible at http://localhost:${port}`);
 });
+
+export { isFirstStartEvent, toggleFirstStartTrue, toggleFirstStartFalse };
