+++ src/events/raw.js
@@ -1,8 +1,5 @@
 import FastLink from "@performanc/fastlink";
-import { EmbedBuilder, ButtonBuilder, ButtonStyle, ActionRowBuilder } from "discord.js";
-import { logger } from "../utils/logger.js";
-import { getVoice, getQueue, autoLeave } from "../services/services.js";
-import { client } from "../musicbot.js";
+import { nowPlaying } from "../services/services.js";
 
 const name = "raw";
 const runOnce = false;
@@ -29,73 +26,9 @@ const lavaClient = FastLink.node.connectNodes(
   },
 );
 
-const overrideChannels = [
-  {
-    guildId: "889971568732684298",
-    channelId: "1139615420400291851",
-  },
-  {
-    guildId: "166740556947390465",
-    channelId: "708165175341088828",
-  },
-];
-
-let isFirstStartEvent = true;
-
-async function nowPlaying(data) {
-  try {
-    const voiceData = await getVoice(data.guildId);
-    const queue = await getQueue(data.guildId);
-    const matchedOverride = overrideChannels.find((override) => override.guildId === data.guildId);
-    const selectedChannelId = matchedOverride ? matchedOverride.channelId : voiceData.channelId;
-    const channel = client.channels.cache.get(selectedChannelId);
-    const queueButton = new ButtonBuilder().setCustomId("queue").setLabel("Show Queue").setStyle(ButtonStyle.Primary);
-    const row = new ActionRowBuilder().addComponents(queueButton);
-
-    if (typeof queue !== "object") {
-      client.user.setPresence({ activities: [{ name: "you sleep", type: 3 }] });
-      autoLeave(data.guildId);
-      return;
-    }
-
-    if (!channel || !channel.isTextBased()) {
-      logger.error(`Channel Id must exist and allow text: ${voiceData.channelId}`);
-      return;
-    }
-
-    const { title, author, uri, artworkUrl } = queue[0].info;
-    const albumName = queue[0].pluginInfo.albumName || "";
-
-    const nowPlaying = new EmbedBuilder()
-      .setTitle(title)
-      .setURL(uri)
-      .setAuthor({
-        name: "Now Playing",
-      })
-      .setDescription(
-        `${author}
-        ${albumName}`,
-      )
-      .setThumbnail(artworkUrl)
-      .setImage("https://raw.githubusercontent.com/nullpat/randy-backend/refs/heads/main/line.png");
-
-    channel.send({ embeds: [nowPlaying], components: [row] });
-    client.user.setPresence({
-      activities: [{ name: `${title} - ${author}`, type: 2 }],
-    });
-  } catch (error) {
-    logger.error(error.stack);
-  }
-}
-
 lavaClient.on("raw", async (data) => {
-  if (data.type === "TrackStartEvent") {
-    if (isFirstStartEvent) {
-      isFirstStartEvent = false;
-      nowPlaying(data);
-    }
-  } else if (data.type === "TrackEndEvent") {
-    nowPlaying(data);
+  if (data.type === "TrackEndEvent") {
+    nowPlaying(data.guildId);
   } else {
     return;
   }
